+++ griptape/drivers/__init__.py
@@ -31,11 +31,12 @@ from .vector.base_vector_store_driver import BaseVectorStoreDriver
 from .vector.local_vector_store_driver import LocalVectorStoreDriver
 from .vector.pinecone_vector_store_driver import PineconeVectorStoreDriver
 from .vector.marqo_vector_store_driver import MarqoVectorStoreDriver
-from .vector.mongodb_vector_store_driver import MongoDbAtlasVectorStoreDriver
+from .vector.mongodb_atlas_vector_store_driver import MongoDbAtlasVectorStoreDriver
 from .vector.redis_vector_store_driver import RedisVectorStoreDriver
 from .vector.opensearch_vector_store_driver import OpenSearchVectorStoreDriver
 from .vector.amazon_opensearch_vector_store_driver import AmazonOpenSearchVectorStoreDriver
 from .vector.pgvector_vector_store_driver import PgVectorVectorStoreDriver
+from .vector.azure_mongodb_vector_store_driver import AzureMongoDbVectorStoreDriver
 
 from .sql.base_sql_driver import BaseSqlDriver
 from .sql.amazon_redshift_sql_driver import AmazonRedshiftSqlDriver
