+++ griptape/drivers/vector/mongodb_atlas_vector_store_driver.py
